Rename broker_test to broker
