+++ augur/broker/broker/routes/__init__.py
@@ -12,7 +12,7 @@ def get_route_files():
 
 def create_all_routes(server):
     for route_file in get_route_files():
-        module = importlib.import_module('.' + route_file, 'broker_test.routes')
+        module = importlib.import_module('.' + route_file, 'broker.routes')
         module.create_routes(server)
 
 def get_file_id(path):
